refactor!: Code movement in FFI crate
