+++ ffi/src/expressions/kernel_visitor.rs
@@ -1,7 +1,5 @@
-//! Defines [`KernelExpressionVisitorState`]. This is a visitor  that can be used by an [`EnginePredicate`]
-//! to convert engine expressions into kernel expressions.
-use std::ffi::c_void;
-
+//! Defines [`KernelExpressionVisitorState`]. This is a visitor that can be used to convert an
+//! engine's native expressions into kernel's [`Expression`] and [`Predicate`] types.
 use crate::{
     AllocateErrorFn, EngineIterator, ExternResult, IntoExternResult, KernelStringSlice,
     ReferenceSet, TryFromStringSlice,
@@ -19,25 +17,6 @@ pub(crate) enum ExpressionOrPredicate {
 #[derive(Default)]
 pub struct KernelExpressionVisitorState {
     inflight_ids: ReferenceSet<ExpressionOrPredicate>,
-}
-
-/// A predicate that can be used to skip data when scanning.
-///
-/// When invoking [`scan::scan`], The engine provides a pointer to the (engine's native) predicate,
-/// along with a visitor function that can be invoked to recursively visit the predicate. This
-/// engine state must be valid until the call to [`scan::scan`] returns. Inside that method, the
-/// kernel allocates visitor state, which becomes the second argument to the predicate visitor
-/// invocation along with the engine-provided predicate pointer. The visitor state is valid for the
-/// lifetime of the predicate visitor invocation. Thanks to this double indirection, engine and
-/// kernel each retain ownership of their respective objects, with no need to coordinate memory
-/// lifetimes with the other.
-///
-/// [`scan::scan`]: crate::scan::scan
-#[repr(C)]
-pub struct EnginePredicate {
-    pub predicate: *mut c_void,
-    pub visitor:
-        extern "C" fn(predicate: *mut c_void, state: &mut KernelExpressionVisitorState) -> usize,
 }
 
 fn wrap_expression(state: &mut KernelExpressionVisitorState, expr: impl Into<Expression>) -> usize {
