+++ ffi/src/expressions/mod.rs
@@ -3,11 +3,31 @@
 use delta_kernel::{Expression, Predicate};
 use delta_kernel_ffi_macros::handle_descriptor;
 
-pub mod engine;
-pub mod kernel;
+use crate::handle::Handle;
+
+pub mod engine_visitor;
+pub mod kernel_visitor;
 
 #[handle_descriptor(target=Expression, mutable=false, sized=true)]
 pub struct SharedExpression;
 
 #[handle_descriptor(target=Predicate, mutable=false, sized=true)]
 pub struct SharedPredicate;
+
+/// Free the memory the passed SharedExpression
+///
+/// # Safety
+/// Engine is responsible for passing a valid SharedExpression
+#[no_mangle]
+pub unsafe extern "C" fn free_kernel_expression(data: Handle<SharedExpression>) {
+    data.drop_handle();
+}
+
+/// Free the memory the passed SharedPredicate
+///
+/// # Safety
+/// Engine is responsible for passing a valid SharedPredicate
+#[no_mangle]
+pub unsafe extern "C" fn free_kernel_predicate(data: Handle<SharedPredicate>) {
+    data.drop_handle();
+}
