+++ ffi/src/scan.rs
@@ -1,6 +1,7 @@
 //! Scan related ffi code
 
 use std::collections::HashMap;
+use std::ffi::c_void;
 use std::sync::{Arc, Mutex};
 
 use delta_kernel::scan::state::{DvInfo, GlobalScanState};
@@ -11,9 +12,7 @@ use delta_kernel_ffi_macros::handle_descriptor;
 use tracing::debug;
 use url::Url;
 
-use crate::expressions::engine::{
-    unwrap_kernel_predicate, EnginePredicate, KernelExpressionVisitorState,
-};
+use crate::expressions::kernel_visitor::{unwrap_kernel_predicate, KernelExpressionVisitorState};
 use crate::expressions::SharedExpression;
 use crate::{
     kernel_string_slice, AllocateStringFn, ExternEngine, ExternResult, IntoExternResult,
@@ -31,6 +30,23 @@ pub struct SharedScan;
 
 #[handle_descriptor(target=ScanMetadata, mutable=false, sized=true)]
 pub struct SharedScanMetadata;
+
+/// A predicate that can be used to skip data when scanning.
+///
+/// When invoking [`scan`], The engine provides a pointer to the (engine's native) predicate,
+/// along with a visitor function that can be invoked to recursively visit the predicate. This
+/// engine state must be valid until the call to [`scan::scan`] returns. Inside that method, the
+/// kernel allocates visitor state, which becomes the second argument to the predicate visitor
+/// invocation along with the engine-provided predicate pointer. The visitor state is valid for the
+/// lifetime of the predicate visitor invocation. Thanks to this double indirection, engine and
+/// kernel each retain ownership of their respective objects, with no need to coordinate memory
+/// lifetimes with the other.
+#[repr(C)]
+pub struct EnginePredicate {
+    pub predicate: *mut c_void,
+    pub visitor:
+        extern "C" fn(predicate: *mut c_void, state: &mut KernelExpressionVisitorState) -> usize,
+}
 
 /// Drop a `SharedScanMetadata`.
 ///
